Tilemap drawing works now.

+++ TaunoKarki/inc/mesh.h
@@ -28,12 +28,12 @@ public:
 	~Mesh();
 
 	const std::vector<Vertex>& getVertices() const { return vertices; }
-	const std::vector<GLushort>& getIndices() const { return indices; }
+	const std::vector<GLuint>& getIndices() const { return indices; }
 private:
-	std::vector<GLushort> indices;
+	std::vector<GLuint> indices;
 	std::vector<Vertex> vertices;
 
-	bool getSimilarVertexIndex(Vertex& vertex, std::map<Vertex, GLushort>& vertexToOutIndex, GLushort& result);
+	bool getSimilarVertexIndex(Vertex& vertex, std::map<Vertex, GLuint>& vertexToOutIndex, GLuint& result);
 	void indexVBO(std::vector<glm::vec3>& vertices, std::vector<glm::vec2>& uvs, std::vector<glm::vec3>& normals);
 };
